Fix: 가격 타입 Long 에서 int 로 변경

+++ costcook/src/main/java/com/costcook/domain/response/RecipeResponse.java
@@ -21,7 +21,7 @@ public class RecipeResponse {
 
 	// Recipe -> response 변환
 	// 전체 목록
-	public static RecipeResponse toDTO(Recipe recipe, double avgRatings, int reviewCount, Long totalPrice) {		
+	public static RecipeResponse toDTO(Recipe recipe, double avgRatings, int reviewCount, int totalPrice) {		
 		// 평점 소수점 첫째자리까지 반올림
 		double roundedAvgRatings = Math.round(avgRatings * 10) / 10.0;
 		
@@ -35,7 +35,7 @@ public class RecipeResponse {
 				.createdAt(recipe.getCreatedAt().format(DateTimeFormatter.ofPattern("yyyy.MM.dd")))
 				.updatedAt(recipe.getUpdatedAt().format(DateTimeFormatter.ofPattern("yyyy.MM.dd")))
 				.servings(recipe.getServings())
-				.price(totalPrice.intValue())
+				.price(totalPrice)
 				.viewCount(recipe.getViewCount())
 				.reviewCount(reviewCount)
 				.avgRatings(roundedAvgRatings)
